fixed the /who all client crash.

+++ world/clientlist.cpp
@@ -669,7 +669,7 @@ void ClientList::SendWhoAll(uint32 fromid,const char* to, int16 admin, Who_All_S
 		}
 
 		// This is the packet header data.
-		uint16 plid = fromid;
+		uint32 plid = fromid;
 		uint16 playerineqstring = WHOALL_PLAYERS;
 		const char line2[] = "---------------------------";
 		uint8 unknown35 = 0x0A;
@@ -700,8 +700,8 @@ void ClientList::SendWhoAll(uint32 fromid,const char* to, int16 admin, Who_All_S
 		uchar *buffer = pack2->pBuffer;
 		uchar *bufptr = buffer;
 
-		memcpy(bufptr, &plid, sizeof(uint16));
-		bufptr += sizeof(uint16);
+		memcpy(bufptr, &plid, sizeof(uint32));
+		bufptr += sizeof(uint32);
 		memcpy(bufptr, &playerineqstring, sizeof(uint16));
 		bufptr += sizeof(uint16);
 		memcpy(bufptr, &line2, strlen(line2));
